Silence warning

+++ src/program.h
@@ -1094,7 +1094,7 @@ void make_program_executable(struct program *p);
 static inline int PIKE_UNUSED_ATTRIBUTE FIND_LFUN(struct program * p, enum LFUN lfun) {
 #ifdef PIKE_DEBUG
     dmalloc_touch(struct program*, p);
-    if (lfun < 0) return find_lfun_fatal(p, lfun);
+    if ((int)lfun < 0) return find_lfun_fatal(p, lfun);
 #endif
     if (p->flags & PROGRAM_FIXED && lfun < NUM_LFUNS) return p->lfuns[lfun];
     return low_find_lfun(p, lfun);
